+++ index.js
@@ -1,13 +1,39 @@
 import * as http from 'http';
 import App from "./src/App.js";
+import * as winston from "winston";
+import * as util from "util";
+
+function transform(info, opts) {
+    const args = info[Symbol.for('splat')];
+    if (args) { info.message = util.format(info.message, ...args); }
+    return info;
+  }
+  
+function utilFormatter() { return {transform}; }
+
+const logger = winston.createLogger({
+    level: process.env.DEBUG_LEVEL || "info",
+    format: winston.format.combine(
+        winston.format.timestamp({format: 'YYYY-MM-DD HH:mm:ss.SSS'}),
+        utilFormatter(),
+        winston.format.colorize(),
+        winston.format.printf(({level, message, label, timestamp}) => `${timestamp} ${label || '-'} ${level}: ${message}`),
+      ),
+      transports: [
+        new winston.transports.Stream({
+          stream: process.stderr,
+          level: process.env.DEBUG_LEVEL || "info",
+        })
+      ],
+});
 
 const app = new App();
 
 const ip = await getPublicIp();
-console.log("===> Got current public ip as ===>", ip);
+logger.info("Got public IP as %s", ip);
 
 const zoneId = await getZoneId();
-console.log("===> Got zone id as ===>", zoneId);
+logger.debug("Got zone ID as %s", zoneId);
 //update();