+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 + Add Docker
 + Get public IP
 + Promisify getZoneId
++ Add Winston logger
 
 ## 0.1.0
 + First version - forked